% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{linreg}
\alias{linreg}
\title{linreg}
\source{
Bates, D. Matrix decompositions for regression analysis. Retrieved from \href{https://pages.stat.wisc.edu/~st849-1/lectures/Orthogonal.pdf}{Department of Statistics at University of Wisconsin–Madison}.
}
\usage{
linreg(formula, data)
}
\arguments{
\item{formula}{A formula object describing how the designmatrix is supposed to look and what the variable of interest is.}

\item{data}{A dataframe where all columns included are numeric}
}
\value{
Returns a regressionobject that has special methods for plot, resid, print, pred, coef and summary
}
\description{
This function is an R implementation of least squares linear regression using the qr decomposition.
}
\examples{
linreg(Sepal.Length ~ Sepal.Width + Petal.Length, iris)

linreg(DAX ~ SMI + CAC, EuStockMarkets)


}
