% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotreg.R
\name{plot.reg}
\alias{plot.reg}
\title{plot.reg}
\source{
See the R documentation for \code{\link[stats]{plot.lm}} for more information about the residuals vs fitted values and Scale-Location.
}
\usage{
\method{plot}{reg}(x, ...)
}
\arguments{
\item{x}{A reg class object}

\item{...}{Other arguments}
}
\value{
Returns a residual analysis which consists of two plots, fitted values vs residual and scale-location for each observation (fitted values vs standardized residual).
}
\description{
This method function creates a residual analysis for reg objects.
}
\examples{
plot(linreg(Sepal.Length ~ Sepal.Width, iris))
}
