% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printreg.R
\name{print.reg}
\alias{print.reg}
\title{print.reg}
\usage{
\method{print}{reg}(x, ...)
}
\arguments{
\item{x}{A reg class object}

\item{...}{other functions}
}
\value{
Returns the coefficients of the linear regression model
}
\description{
This prints out a regression model created from the function linreg.
}
\examples{
print(linreg(Sepal.Length ~ Sepal.Width, iris)) ## Coefficients: 6.526223  -0.2233611
}
