% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\docType{class}
\name{reg-class}
\alias{reg-class}
\alias{reg}
\title{linreg}
\source{
Dijkstra's algorithm. (September, 2025). Retrieved from \href{https://en.wikipedia.org/wiki/QR_decomposition}{Wikipedia}.
}
\arguments{
\item{formula}{A formula object describing how the designmatrix is supposed to look and what the variable of interest is.}

\item{data}{A dataframe where all columns included are numeric}
}
\value{
Returns a regressionobject that has special methods for plot, resid, pred, coef, summary
}
\description{
This function is an R implementation of least squares linear regression using the qr decomposition.
}

\examples{
linreg(Sepal.Length ~ Sepal.Width + Petal.Length, iris)

linreg(DAX ~ SMI + CAC, EuStockMarkets)


}
