% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryreg.R
\name{summary.reg}
\alias{summary.reg}
\title{resid.reg}
\usage{
\method{summary}{reg}(x)
}
\arguments{
\item{x}{A reg class object}
}
\value{
Returns the beta-coefficients from the model as well as their standard errors and corresponding t- and p-values.
}
\description{
resid.reg
}
\examples{
summary(linreg(Sepal.Length ~ Sepal.Width, iris))
}
