% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefreg.R
\name{coef.reg}
\alias{coef.reg}
\title{coef.reg}
\usage{
\method{coef}{reg}(object, ...)
}
\arguments{
\item{object}{A reg class object}

\item{...}{other arguments}
}
\value{
Returns all estimated coefficients for the linear model
}
\description{
coef.reg
}
\examples{
coef(linreg(Sepal.Length ~ Sepal.Width, iris)) ## 6.526223  -0.2233611
}
