% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotreg.R
\name{plot.reg}
\alias{plot.reg}
\title{plot.reg}
\usage{
\method{plot}{reg}(x)
}
\arguments{
\item{x}{A reg class object}
}
\value{
Returns a residual analysis which consists of two plots, fitted values vs residual and scale-location for each observation (fitted values vs standardized residual).
}
\description{
plot.reg
}
\examples{
plot(linreg(Sepal.Length ~ Sepal.Width, iris))
}
