% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predreg.R
\name{pred}
\alias{pred}
\title{Alternative prediction function}
\usage{
pred(object, ...)
}
\arguments{
\item{object}{model object to pass to predict}

\item{...}{additional arguments to pass to predict}
}
\description{
See \code{\link[stats]{predict}} for details.
}
