% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predreg.R
\name{pred}
\alias{pred}
\title{Alternative prediction function}
\usage{
pred(object, ...)
}
\arguments{
\item{object}{object that is used for the predictions}

\item{...}{other arguments}
}
\description{
This is an alternative function for predict. See \code{\link[stats]{predict}} for details.
}
