% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predreg.R
\name{predict.reg}
\alias{predict.reg}
\title{predict.reg}
\usage{
\method{predict}{reg}(model, new_data = FALSE)
}
\arguments{
\item{model}{A reg class object}

\item{new_data}{a data.frame with character and numerical values. If no data.frame is given, the default to return is the fitted values for the training data.}
}
\value{
Predicted values for the new_data.
}
\description{
predict.reg
}
\examples{
predict(linreg(Sepal.Length ~ Sepal.Width, iris[-c(1:5), ]), iris[1:5, ])
}
