% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residreg.R
\name{residuals.reg}
\alias{residuals.reg}
\title{resid.reg}
\usage{
\method{residuals}{reg}(object, ...)
}
\arguments{
\item{object}{A reg class object}

\item{...}{other arguments}
}
\value{
Returns all the residuals for the linear model
}
\description{
resid.reg
}
\examples{
residuals(linreg(Sepal.Length ~ Sepal.Width, iris))[1:5]
}
